refactor: 테스트 파일을 별도의 폴더로 분리 / 도메인 별로 핸들러 분리
