+++ src/mocks/handlers/index.ts
@@ -1,0 +1,3 @@
+import { productsHandlers } from './products';
+
+export const handlers = [...productsHandlers];
