+++ tsconfig.json
@@ -20,6 +20,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src", "__test__/products.test.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
